Add --stdin-filepath
Before this change users who passed --stdin could not customize the
configuration of prettier-standard.

After this change, --stdin-filepath can be used to load configuration
as though the --stdin input came from a given file.

This is necessary for editor plugins like ale [1] to provide
configuration. They cannot call prettier-standard on the actual file
as it is managed by the editor.

An alternative workaround would be to allow passing a path to a config
file directly, but this approach is both consistent with prettier, and
backwards compatible in the face of any changes to how configuration
files are resolved.

[1] https://github.com/dense-analysis/ale